+++ app/javascript/vanilla/initializers/stickyNavbar.js
@@ -1,6 +1,12 @@
-function getPositionRelativeToElement(element) {
+function getRelativeElement(element) {
   const selector = element.getAttribute('data-relative-to')
   const target = selector && document.querySelector(selector)
+
+  return document.querySelector(selector)
+}
+
+function getPositionRelativeToElement(element) {
+  const target = getRelativeElement(element)
 
   return target ? target.getBoundingClientRect() : {}
 }
@@ -27,15 +33,17 @@ function removeSticky(element) {
 export function setStickyNavbar(element) {
   const handleScroll = () => {
     const { scrollY } = window
+    const { parentElement } = element
+    const offsetTop = (getRelativeElement(element) || parentElement).offsetTop
 
-    if (scrollY > element.parentElement.offsetTop) {
+    if (scrollY > offsetTop) {
       setSticky(element)
     } else {
       removeSticky(element)
     }
 
-    element.parentElement.style.minHeight = `${element.clientHeight}px`
-    element.style.width = `${element.parentElement.clientWidth}px`
+    parentElement.style.minHeight = `${element.clientHeight}px`
+    element.style.width = `${parentElement.clientWidth}px`
   }
 
   const resizeNavbar = () => {